+++ app/assets/scripts/minireel/editor.js
@@ -43,9 +43,9 @@ VideoCardController           , c6embed) {
         }])
 
         .service('EditorService', ['MiniReelService','$q','c6AsyncQueue','VoteService',
-                                   'CollateralService','SettingsService','ThumbnailService',
+                                   'CollateralService','SettingsService',
         function                  ( MiniReelService , $q , c6AsyncQueue , VoteService ,
-                                    CollateralService , SettingsService , ThumbnailService ) {
+                                    CollateralService , SettingsService ) {
             var _private = {},
                 queue = c6AsyncQueue(),
                 beforeSyncFns = {};
@@ -165,21 +165,6 @@ VideoCardController           , c6embed) {
                     });
                 }
 
-                function fetchThumbs(proxy) {
-                    return $q.all(proxy.data.deck.filter(function(card) {
-                        return (/^(yahoo|aol|rumble)$/).test(card.data.service) && !card.thumb;
-                    }).map(function(card) {
-                        var service = card.data.service,
-                            videoid = card.data.videoid;
-
-                        return ThumbnailService.getThumbsFor(service, videoid)
-                            .ensureFulfillment()
-                            .then(function(thumb) {
-                                card.thumb = thumb.large;
-                            });
-                    }));
-                }
-
                 function beforeSync(proxy) {
                     return $q.all(Object.keys(beforeSyncFns).map(function(id) {
                         return beforeSyncFns[id](proxy);
@@ -190,7 +175,7 @@ VideoCardController           , c6embed) {
 
                 return beforeSync(proxy)
                     .then(function() {
-                        return $q.all([syncWithCollateral, fetchThumbs].map(function(fn) {
+                        return $q.all([syncWithCollateral].map(function(fn) {
                             return fn(proxy);
                         }));
                     })
@@ -1220,7 +1205,7 @@ VideoCardController           , c6embed) {
                     };
 
                     this.afterModel = function(model) {
-                        var types = ['image', 'video', 'videoBallot', 'text', 'instagram'];
+                        var types = ['image', 'video', 'text', 'instagram'];
 
                         if (!model || types.indexOf(model.type) < 0 || model.sponsored) {
                             c6State.goTo('MR:Editor', null, {}, true);
@@ -1249,22 +1234,6 @@ VideoCardController           , c6embed) {
                         }
 
                         return card;
-                    };
-                }])
-
-                .state('MR:EditCard.Ballot', ['c6UrlMaker','MiniReelService',
-                function                     ( c6UrlMaker , MiniReelService ) {
-                    this.controller = 'GenericController';
-                    this.controllerAs = 'EditCardBallotCtrl';
-                    this.templateUrl = 'views/minireel/editor/edit_card/ballot.html';
-                    this.model = function() {
-                        var card = this.cParent.cModel;
-
-                        if (card.type !== 'videoBallot') {
-                            MiniReelService.setCardType(card, 'videoBallot');
-                        }
-
-                        return card.data.ballot;
                     };
                 }])
 
@@ -1333,7 +1302,6 @@ VideoCardController           , c6embed) {
                         switch (this.model.type) {
                         case 'image':
                         case 'video':
-                        case 'videoBallot':
                             return [this.copyComplete].indexOf(false) < 0 &&
                                 !EditorCtrl.errorForCard(this.model);
 
@@ -1350,7 +1318,6 @@ VideoCardController           , c6embed) {
                         switch (this.model.type) {
                         case 'image':
                         case 'video':
-                        case 'videoBallot':
                             return ['title'].map(function(prop) {
                                 return !!model[prop];
                             }).indexOf(false) < 0;
@@ -1373,8 +1340,7 @@ VideoCardController           , c6embed) {
                         case 'instagram':
                             return !!model.data.id;
                         case 'video':
-                        case 'videoBallot':
-                            return true;
+                            return !!(model.data.service && model.data.videoid);
                         default:
                             return undefined;
                         }
@@ -1437,15 +1403,6 @@ VideoCardController           , c6embed) {
                         icon: 'play',
                         required: true
                     },
-                    ballotTab = {
-                        name: 'Questionnaire',
-                        sref: 'MR:EditCard.Ballot',
-                        icon: 'ballot',
-                        required: false,
-                        customRequiredText: [
-                            'Indicates required field (to include a questionnaire)'
-                        ].join('')
-                    },
                     imageTab = {
                         name: 'Image Content',
                         sref: 'MR:EditCard.Image',
@@ -1465,9 +1422,8 @@ VideoCardController           , c6embed) {
                     case 'instagram':
                         return [instagramTab];
                     case 'video':
-                    case 'videoBallot':
                     case 'text':
-                        return [videoTab, copyTab, ballotTab];
+                        return [videoTab, copyTab];
                     default:
                         return [];
                     }
@@ -1481,21 +1437,7 @@ VideoCardController           , c6embed) {
                 }
             };
 
-            this.setIdealType = function() {
-                var choices;
-
-                if (!(/^video/).test(this.model.type)) { return; }
-
-                if (!this.model.data.videoid) {
-                    return MiniReelService.setCardType(this.model, 'text');
-                }
-
-                choices = (this.model.data.ballot || {}).choices || [];
-
-                if ((!choices[0] || !choices[1]) && this.model.type !== 'video') {
-                    MiniReelService.setCardType(this.model, 'video');
-                }
-            };
+            this.setIdealType = function() {};
 
             this.save = function() {
                 var deck = EditorCtrl.model.data.deck,
@@ -1685,7 +1627,7 @@ VideoCardController           , c6embed) {
                 };
 
                 this.model = function() {
-                    return MiniReelService.createCard('videoBallot');
+                    return MiniReelService.createCard('video');
                 };
 
                 this.afterModel = function(card, params) {
@@ -1705,7 +1647,7 @@ VideoCardController           , c6embed) {
             var EditorCtrl = $scope.EditorCtrl,
                 minireel = EditorCtrl.model;
 
-            this.type = 'videoBallot';
+            this.type = 'video';
 
             this.edit = function() {
                 var card = MiniReelService.setCardType(this.model, this.type),